profile settings break into components
to be less chunkyy
